Explicitly cast degree to avoid implicit conversion
The cast is save, because the rhs polynomial is checked for having at
least degree zero and the lhs polynomial is required to have a degree of
greater than the degree of the rhs polynomial, thus greater than 0.
Therefore, the difference of the degrees is always positive.

+++ src/gf/polynomial.h
@@ -17,7 +17,8 @@ class polynomial : public std::vector<Element> {
     if (lhs.degree() < rhs.degree())
       return std::make_tuple(polynomial(1, GF::zero), lhs);
 
-    const size_t quotient_degree = lhs.degree() - rhs.degree();
+    const size_t quotient_degree =
+        static_cast<size_t>(lhs.degree() - rhs.degree());
     polynomial q(quotient_degree + 1, GF::zero);
     polynomial r(lhs);
